projeto_chamados/requirements.txt
asgiref==3.11.0
build==1.3.0
click==8.3.1
Django==6.0
django-role-permissions==3.2.0
packaging==25.0
pip-tools==7.5.2
pyproject_hooks==1.2.0
setuptools==80.9.0
sqlparse==0.5.4
wheel==0.45.1
